% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\docType{data}
\name{fx}
\alias{fx}
\title{minimum variance hedge ratio function}
\format{
A data frame with 5047 rows and 2 variables:
\describe{
\item{EURO/USD}{exchange rate}
\item{EURO futures price}{EURO futures price data}
...
}
}
\source{
\url{http://www.bloomberg.com/}
}
\usage{
fx
}
\arguments{
\item{x}{(n x 2) matrix of input that consists of spot price in the first column and futures prices in the second column.}

\item{WinLen}{Length of window}
}
\value{
(list(HR = HR, HE = HE))
\item{HR}{Length(nWin x 1) optimal hedge ratio matrix}
\item{HE}{Length(nWin x 1) hedging effectiveness matrix}
}
\description{
minimum variance hedge ratio function
}
\keyword{datasets}
