% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\name{mvhr}
\alias{mvhr}
\title{minimum variance hedge ratio function}
\usage{
mvhr(x, WinLen)
}
\arguments{
\item{x}{(n x 2) matrix of input that consists of spot price in the first column and futures prices in the second column.}

\item{WinLen}{Length of window}
}
\value{
(list(HR = HR, HE = HE))
\item{HR}{Length(nWin x 1) optimal hedge ratio matrix}
\item{HE}{Length(nWin x 1) hedging effectiveness matrix}
}
\description{
minimum variance hedge ratio function
}
\examples{
data(fx)
spot_price = fx$...1
futures_price = fx$...2
x = cbind(spot_price, futures_price)
Winlen <- 15
mvhr(x, Winlen)
}
