% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\name{semivar}
\alias{semivar}
\title{Semivariance hedge ratio}
\usage{
semivar(v, w)
}
\arguments{
\item{v}{length(WinLen x 1) price vector}

\item{w}{length(WinLen x 1) column vector for weight}
}
\value{
semivariance of the input vector, returns a scalar
}
\description{
This function will calculate the single commodity hedge ratio by minimizing the semi variance of the returns.
}
\examples{
WinLen=15
w <- (matrix(1, WinLen, 1)) / WinLen
v <- matrix(runif(WinLen, min=0, max=10), WinLen, 1)
semivar(v,w) #return scalar
}
