% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\docType{data}
\name{fx}
\alias{fx}
\title{EURO/USD exchange rate and futures price of EURO.
A dataset containing the prices and other attributes of almost 20 years}
\format{
A data frame with 5047 rows and 2 variables:
\describe{
\item{...1}{EURO/USD exchange rate}
\item{...2}{EURO futures price data}
}
}
\source{
\url{http://www.bloomberg.com/}
}
\usage{
data(fx)
WinLen <-15
x <- fx
}
\description{
EURO/USD exchange rate and futures price of EURO.
A dataset containing the prices and other attributes of almost 20 years
}
\keyword{datasets}
